hotel_booking_full_complete/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
python-dotenv
django-cors-headers
drf-yasg
celery
redis
